CSS-9328 stream connector (#1344)
* feat: stream proxier

* Move apiproxy into a separate file

* test fixes and pr comments

* Add godoc

* fix tests and use new error code for session token login

* fix cmd tests

Juju changed their CLI login in https://github.com/juju/juju/pull/17667 to fix a bug so now CLI commands check that the user in the accounts.yaml file you were trying to login as, matches the user you ended up logging in as.
In JIMM's tests, we have an in-memory accounts struct that was empty so this was causing an issue.

* Test comment

* Further test fixes

* linting fixes

* address comments

* fix linting errors

* address comments

* Fixes after merging v3

* revert mechanical changes

- Revert creating an apiproxy.go file and save that for a follow up to avoid creating large diffs unnecessarily.

* break up large ServeWS function

* fix missing body close

* pr comments

- use specific routes
- more logging

* moved stream proxy into rpc package

* PR comments

- Add GetModel method on JIMM.
- Added comment about why no debug logs are present

* PR comments

- Add further docstring
- Reorder some code