fix(bootstrap-lock): fix bootstrap lock by using a table instead of advisory lock
Before we were using a session advisory lock to implement a distribuited lock. This has some issue
because you can lock multiple times if you are re-using the same session, and you need to unlock it
from the same connection that locked it. So, we've switched to a simpler implementation using a
table and a transaction with SELECT FOR UPDATE.